+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.2.9](https://github.com/rudashi/stringable/compare/v1.2.8...v1.2.9) (2022-10-15)
+
+### Feature
+* Add `toInteger` method
+* Add `toFloat` method
+* Add `toBoolean` method
+* Add `toDate` method
+
 ## [1.2.8](https://github.com/rudashi/stringable/compare/v1.2.7...v1.2.8) (2022-10-15)
 
 ### Feature